requirements.txt
asposestorage==1.0.2
brisque==0.0.16
config==0.5.1
core==1.0.1
dataset==1.6.2
diffusers==0.31.0
huggingface_hub==0.26.2
imageio==2.36.0
lpips==0.1.4
matplotlib==3.10.3
model==0.6.0
numpy==2.2.6
omegaconf==2.3.0
openai==1.82.1
opencv_contrib_python==4.11.0.86
pandas==2.2.3
Pillow==11.2.1
regex==2024.11.6
SAM_2==1.0
scipy==1.15.3
skimage==0.0
supervision==0.25.1
torch==2.5.1
torchmetrics==1.6.0
torchvision==0.20.1
tqdm==4.66.5
transformers==4.47.1
utils==1.0.2
